Move image handling to the correct package
